requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    bazel run //:requirements.update
#
absl-py==2.0.0 \
    --hash=sha256:9a28abb62774ae4e8edbe2dd4c49ffcd45a6a848952a5eccc6a49f3f0fc1e2f3 \
    --hash=sha256:d9690211c5fcfefcdd1a45470ac2b5c5acd45241c3af71eed96bc5441746c0d5
    # via -r requirements.txt
